pyproject.toml
[tool.poetry]
name = "TODO-project-name"
version = "0.0.1"
description = ""
authors = ["Thorne Wolf <thornecanyon@gmail.com>"]
readme = "README.md"
packages = [{include = "project"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-executor = "^1.0.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
gunicorn = "^20.1.0"
flask-migrate = "^4.0.4"
fastapi = {extras = ["all"], version = "^0.95.0"}
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.7"
beautifulsoup4 = "^4.12.0"
cachetools = "^5.3.0"
absl-py = "^1.4.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
